% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Title}
\usage{
draw_names(NameVec)
}
\arguments{
\item{NameVec}{give a vector with names}
}
\value{
plot simultaneously the popularity of more than one name with ggplot
}
\description{
Title
}
\examples{
draw_names (c("Diane", "Vincent"))
}
