% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction my csv reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{Title}
\usage{
my_csv_reader(folder)
}
\arguments{
\item{folder}{Put the name of a folder}
}
\value{
Takes all the csv from a folder Import them and put them in a list Returns the list
}
\description{
Title
}
