% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_multi.R
\name{read_excel_multi}
\alias{read_excel_multi}
\title{Title}
\usage{
read_excel_multi(file)
}
\arguments{
\item{file}{enter a file name}
}
\value{
import ALL sheets from an xlsx file
}
\description{
Title
}
\examples{
\dontrun{
read_excel_multi("titanic_train.csv")
}
}
