% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Title}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{dataset}{Give a data set}

\item{filename}{Give its file names}

\item{row.names}{Ne prend pas le nom des lignes}

\item{...}{other parameters}
}
\value{
saves a Data.frame in a csv file and also outputs the full path of where the csv has been saved.
}
\description{
Title
}
\examples{

}
