% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonction draw a name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Title}
\usage{
draw_a_name(the_name = n, the_sex = s)
}
\arguments{
\item{the_name}{: give a name}

\item{the_sex}{: give the sex}
}
\value{
which can plot the popularity of babynames (using prenoms) in times use the_name and the_sex as arguments.
}
\description{
Title
}
\examples{
draw_a_name(the_name= "Diane", the_sex = "F")
}
